// src/services/events/event.enums.ts
export enum AudioEvent {
  Audio = 'audio',
  Play = 'play',
  Pause = 'pause',
  PlayToggle = 'play-toggle',
  VolumeChange = 'volume-change',
  FadeInStart = 'fade-in-start',
  FadeIn = 'fade-in',
  FadeInEnd = 'fade-in-end',
  FadeOutStart = 'fade-out-start',
  FadeOut = 'fade-out',
  FadeOutEnd = 'fade-out-end',
  Loop = 'loop',
  Restart = 'restart',
  Move = 'move',
  MoveUp = 'move-up',
  MoveDown = 'move-down',
  DeletePlayer = 'delete-player',
  DeletePlaylist = 'delete-playlist',
  RenamePlaylist = 'rename-playlist',
  AddPlaylist = 'add-playlist',
  AddPlayer = 'add-player',
  Save = 'save',
  Clear = 'clear',
  FadeInPlaylist = 'fade-in-playlist',
  FadeOutPlaylist = 'fade-out-playlist',
  DownloadFile = 'download-file',
  UploadFile = 'upload-file',
  ApiKeyInput = 'api-key',
  LoadPlayers = 'load-players',
}